refactor(poster): Create a base poster abstraction
