+++ src/handlers/poster-base/index.tsx
@@ -1,10 +1,24 @@
-import './index.scss';
-import { useEffect, useState } from 'react';
-import { FooterSize, getPosters, Poster } from '../../api';
-import ProgressBar from './components/ProgressBar';
+import '../poster-gewis/components/index.scss';
+import { ReactNode, useEffect, useState } from 'react';
+import { getPosters, Poster } from '../../api';
 import PosterCarousel from './components/Carousel';
 
-export default function PosterView() {
+export interface OverlayProps {
+  poster: Poster;
+  posterTitle: string;
+  seconds: number;
+  posterIndex: number;
+  nextPoster: () => void;
+  pausePoster: () => void;
+  borrelMode?: boolean;
+}
+
+interface Props {
+  overlay: (overlayProps: OverlayProps) => ReactNode;
+  localPosterRenderer?: (poster: Poster, visible: boolean, setTitle?: (title: string) => void) => ReactNode;
+}
+
+export default function PosterBaseView({ overlay, localPosterRenderer }: Props) {
   const [posters, setPosters] = useState<Poster[]>();
   const [borrelMode, setBorrelMode] = useState(false);
   const [posterIndex, setPosterIndex] = useState<number>();
@@ -76,17 +90,20 @@ export default function PosterView() {
       style={{ backgroundImage: 'url("base/poster-background.png")' }}
     >
       <div className="overflow-hidden w-full h-full">
-        <PosterCarousel posters={posters || []} currentPoster={!posterIndex ? 0 : posterIndex} setTitle={setTitle} />
-        <ProgressBar
-          title={title}
-          seconds={posterTimeout !== undefined ? selectedPoster?.timeout : undefined}
-          posterIndex={posterIndex}
-          minimal={selectedPoster?.footer === FooterSize.MINIMAL}
-          hide={selectedPoster?.footer === FooterSize.HIDDEN}
-          borrelMode={borrelMode}
-          nextPoster={nextPoster}
-          pausePoster={pausePoster}
+        <PosterCarousel
+          posters={posters || []}
+          currentPoster={posterIndex < 0 ? 0 : posterIndex}
+          setTitle={setTitle}
+          localPosterRenderer={localPosterRenderer}
         />
+        {overlay({
+          poster: selectedPoster,
+          posterTitle: title,
+          seconds: posterTimeout !== undefined ? selectedPoster?.timeout : undefined,
+          posterIndex: posterIndex,
+          nextPoster: nextPoster,
+          pausePoster: pausePoster,
+        })}
       </div>
     </div>
   );
