+++ src/handlers/poster-gewis/components/types/BorrelPriceListPoster.tsx
