+++ src/handlers/poster-gewis/components/types/BorrelWallOfShame.tsx
