+++ src/handlers/poster-gewis/components/types/LogoPoster.tsx
