+++ src/handlers/poster-gewis/components/types/ScrollAnimation.scss
