+++ src/handlers/poster-gewis/components/types/TrainPoster.tsx
